User login and session service

+++ src/main/java/org/sem/database/DAOInterface.java
@@ -1,12 +1,11 @@
 package org.sem.database;
 
-import java.sql.SQLException;
 import java.util.List;
 import java.util.Optional;
 
 public interface DAOInterface<T> {
     // Optional => use custom type Class => exists data Class(id, name), => not exists null object
-    Optional<T> get(Integer id);
+    Optional<T> get(Long id);
     List<T> getAll();
     T save(T t);
     Boolean delete(T t);
